Fix removeOldFiles (which didn't work at all) and respective test.
We now make sure to remove the oldest files and not to remove the currently open
file. The test was broken because the second roll would cause a rename to
an existing file so the number of files would not increase independently
of whether there was a rotateCount or not.